chore: move MapPath to ObjectMapping
